requirements.txt
SQLAlchemy==2.0.27
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
plotly==5.19.0
altair==5.3.0  
mysqlclient==2.1.0
matplotlib==3.8.4
fpdf2==2.7.9
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==   0.4.0
chromedriver-autoinstaller==0.6.4
kaleido== 0.2.1
selenium== 4.23.1
vl-convert-python==1.5.1
openpyxl
folium == 0.17.0
pydeck==0.8.1b0
<<<<<<< HEAD
streamlit_folium == 0.22.0
=======
streamlit_folium == 0.22.0
mysql-connector-python
streamlit == 1.32.2
>>>>>>> f2ad9b042b534aa2d0305a18dbd50d25994a5f14
